Further progress for peptides Restructure subcommand
