+++ src/graph/peptide.rs
@@ -1,0 +1,5 @@
+use crate::translation::amino_acids::AminoAcid;
+
+pub(crate) struct Peptide {
+    pub(crate) sequence: Vec<AminoAcid>,
+}
